create community select ticker from dropdownWithSearch
