+++ src/components/common/dropdownWithSearch/index.scss
@@ -1,0 +1,117 @@
+@import "/src/styles";
+.dropdown_with_search {
+  box-sizing: border-box;
+  background-color: var(--secondary-bg);
+  .search_input {
+    width: 50%;
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    border-radius: 8px;
+    background-color: var(--tertiary-bg);
+    padding: 8px;
+    @include lessThan(md) {
+      width: 100%;
+    }
+    .s_icon {
+      width: 24px;
+      height: 24px;
+      color: #757575;
+    }
+  }
+  .custom_dropdown {
+    position: absolute;
+    width: calc(100% - 52%);
+    height: 300px;
+    overflow-y: scroll;
+    padding: 16px 8px 16px 16px;
+    background-color: var(--secondary-bg);
+    box-shadow: var(--card-shadow);
+    z-index: 100;
+    border-radius: 16px;
+    margin-top: 0.5rem;
+    // padding: 0.5rem 0 0.5rem 0.5rem;
+    @include lessThan(md) {
+      width: 100%;
+    }
+  }
+
+  .dropdown_input {
+    cursor: pointer;
+    border: none !important;
+    height: 25px;
+    width: -moz-fit-content;
+    width: 100%;
+    padding: 0 5px !important;
+    background-color: transparent !important;
+    color: var(--black);
+    font-size: $font-size-sm;
+    outline: none;
+    &:focus {
+      border: none !important;
+    }
+    @include lessThan(md) {
+      font-size: $font-size-xsm;
+      padding: 0;
+    }
+  }
+
+  .dropdown_list {
+    border: 0.25rem;
+  }
+
+  .ant-list {
+    max-height: 360px;
+    overflow-y: scroll;
+  }
+
+  .dropdown_list .ant-list-item {
+    cursor: pointer;
+    padding: 8px 8px !important;
+    border-radius: 8px;
+    color: var(--body-bg);
+  }
+
+  .c_post {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    flex-direction: row;
+    // margin-bottom: 1rem;
+    gap: 10px;
+    align-items: center;
+    // height: 30px;
+    img {
+      width: 30px;
+      height: 30px;
+      border-radius: 50%;
+      margin-right: 8px;
+    }
+    .u_bx {
+      width: 100%;
+      display: flex;
+      justify-content: space-between;
+      // line-height: 15px;
+      .u_txt {
+        font-size: $font-size-md;
+
+        color: var(--primary-text);
+        @include lessThan(sm) {
+          font-size: $font-size-sm;
+        }
+      }
+      .u_follow {
+        font-size: $font-size-sm;
+
+        color: var(--label-text);
+        @include lessThan(sm) {
+          font-size: $font-size-xsm;
+        }
+      }
+    }
+  }
+
+  .dropdown_list .ant-list-item:hover {
+    background-color: var(--tertiary-bg);
+  }
+}
